Implemented Voting_model simulations
Potential bug identified: Due to the construction and implementation of laziness, with 0.75 laziness, the voters are voting on the same 25% of projects (Projects 99-76). With expertise 1.0, the variance of people who actually vote is low (as every voter votes “ideal funding”), thus the difference 0.

Future Action: We need to randomize the projects voters vote on when laziness is not 0.